u32 -> i64

+++ src/day1.rs
@@ -1,11 +1,11 @@
 use std::io;
 use crate::input;
 
-fn part1(data: &[u32]) -> usize {
+fn part1(data: &[i64]) -> usize {
     data.windows(2).filter(|x| x[1] > x[0]).count()
 }
 
-fn part2(data: &[u32]) -> usize {
+fn part2(data: &[i64]) -> usize {
     data.windows(4).filter(|w| w[3] > w[0]).count()
 }
